Extract stdin prospector
This a first step in the direction of having a package for each prospector together with its harvester. Stdin prospector was tackled first as it has much less dependencies. Until the refactoring is completed there will be some duplicated code.
